requirements.txt
einops==0.3.0
fvcore==0.1.5.post20210604
matplotlib==3.3.2
mmcv==1.3.5
numpy==1.18.5
opencv_python==4.5.2.54
pavi==0.0.1
Pillow==9.1.0
portalocker==2.0.0
requests==2.24.0
scikit_image==0.17.2
scipy==1.5.2
setproctitle==1.2.3
skimage==0.0
tifffile==2020.9.3
timm==0.3.2
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.2
